CLEANUP: rename blueview_driver to mil_blueview_driver
All packages in mil_common that are no submodules should be
prefixed with 'mil_'.
