+++ drivers/mil_blueview_driver/README.md
@@ -7,7 +7,7 @@ Currently this ROS driver depends on MIL's catkin-based version of the blue view
 If you are interested in using this outside of MIL, you must ensure this library is
 linked with the BlueView SDK and has the BlueView SDK include folder.
 
-``` rosrun blue_view_driver blue_view_driver _device:=SONARIPADDRESS ```
+``` rosrun mil_blueview_driver blueview_driver _device:=SONARIPADDRESS ```
 
 ## Config
