+++ drivers/mil_blueview_driver/include/blueview_ros_driver.hpp
@@ -1,4 +1,4 @@
-#include <blueview_driver/BlueViewPing.h>
+#include <mil_blueview_driver/BlueViewPing.h>
 #include <bvt_sdk.h>
 #include <cv_bridge/cv_bridge.h>
 #include <image_transport/image_transport.h>
@@ -8,7 +8,7 @@
 #include <functional>
 #include <opencv2/core/core.hpp>
 #include <stdexcept>
-#include "BVWrapper.hpp"
+#include <blueview_wrapper.hpp>
 
 class BlueViewRosDriver
 {
@@ -25,7 +25,7 @@ private:
   image_transport::Publisher grayscale_pub, color_pub;
   ros::Publisher raw_pub;
   cv_bridge::CvImagePtr grayscale_img, color_img;
-  blueview_driver::BlueViewPingPtr ping_msg;
+  mil_blueview_driver::BlueViewPingPtr ping_msg;
   size_t i = 0;
 
   ros::Timer timer;
