+++ drivers/mil_blueview_driver/msg/BlueViewPing.msg
