+++ drivers/mil_blueview_driver/package.xml
@@ -1,7 +1,7 @@
 <package>
-  <name>blueview_driver</name>
+  <name>mil_blueview_driver</name>
   <version>1.0.0</version>
-  <description>Ros driver for Telodyne blue view, depends on blue view library</description>
+  <description>ROS driver for Teledyne blue view, depends on blueview sdk</description>
   <maintainer email="tessbianchi@ufl.edu">Tess Bianchi</maintainer>
   <maintainer email="kma1660@gmail.com">Kevin Allen</maintainer>
